### 2025-06-20 #8442 @editorialbot
[PRE REVIEW]: WAND (Working-memory Active-fatigue with N-back Difficulty): A Modular Software Suite for Cognitive Fatigue Research
**Submitting author:** @brodie-neuro (<a href="http://orcid.org/0009-0002-8466-5423">Brodie Mangan</a>)
**Repository:** https://github.com/brodie-neuro/WAND-practice-and-fatigue-induction
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
