### 2025-06-23 d6a297b
Update paper.bib with missing DOIs
