+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={1285--1289},
   year={2019},
-  organization={SAGE Publications Sage CA: Los Angeles, CA}
+  organization={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={10.1177/1071181319631106}
 }
 
 @article{holroyd2024controllosphere,
@@ -14,7 +15,8 @@
   author={Holroyd, Clay B},
   journal={Psychological Review},
   year={2024},
-  publisher={American Psychological Association}
+  publisher={American Psychological Association},
+  doi={10.31234/osf.io/2qmhg}
 }
 
 @article{hopstaken2015multifaceted,
@@ -25,7 +27,8 @@
   number={3},
   pages={305--315},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/psyp.12339}
 }
 
 @article{peirce2019psychopy2,
@@ -35,7 +38,8 @@
   volume={51},
   pages={195--203},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.3758/s13428-018-01193-y}
 }
 
 @article{pessiglione2025origins,
@@ -43,5 +47,6 @@
   author={Pessiglione, Mathias and Blain, Bastien and Wiehler, Antonius and Naik, Shruti},
   journal={Trends in Cognitive Sciences},
   year={2025},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.tics.2025.02.005}
 }
