### 2025-06-28 d9efe1a
docs: Refine documentation for greater clarity for JOSS review
