### 2025-07-01 #8442 @brodie-neuro
Hi @mstimberg. It is a pleasure to make your acquaintance. Thank you for the update. If you need anything from me, I will respond within 24 hours typically. Best, Brodie
