### 2025-07-30 #8667 @mstimberg
#👋🏼 @brodie-neuro @olebialas @leodutriaux this is the review thread for the paper. All of our communications will happen here from now on.

at the top of a new comment in this thread.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/8667#issue-3276370541).

Please feel free to ping me (@mstimberg) if you have any questions/concerns. Looking forward to working with you!
