### 2025-07-31 #8667 @mstimberg
:wave: Hi everyone, this is just to let you know that I will be travelling for most of August and will be enjoying some indecently long summer vacation – my reaction times here will therefore most likely be longer than usual.
