+++ WAND_full_induction.py
@@ -1,4 +1,32 @@
 #!/usr/bin/env python3
+"""
+WAND — Full Fatigue Induction
+
+Full cognitive-fatigue induction protocol using the WAND
+(Working-memory Adaptive-fatigue with N-back Difficulty) model.
+
+Participants complete adaptive Sequential, Spatial, and Dual N-back tasks to
+induce active mental fatigue, with detailed behavioural performance and
+subjective measures collected throughout.
+
+Designed for cognitive-fatigue research (EEG-compatible and behavioural-only).
+
+Author
+------
+Brodie E. Mangan
+
+Version
+-------
+1.0
+
+Environment
+-----------
+Tested on Windows, Python 3.8. See requirements.txt for exact pins.
+
+License
+-------
+MIT (see LICENSE).
+"""
 import argparse
 import csv
 import logging
@@ -10,27 +38,13 @@ import time
 
 from psychopy import core, event, visual
 
-"""
-WAND-fatigue-induction
-
-Full cognitive fatigue induction protocol using the WAND (Working-memory Adaptive-fatigue with N-back Difficulty) model.
-
-Participants complete adaptive Sequential, Spatial, and Dual N-back tasks to induce active mental fatigue,
-with detailed behavioural performance and subjective measures collected throughout.
-
-Designed for cognitive fatigue research, including EEG or behavioural-only implementations.
-
-Requires: PsychoPy, Python 3.8+.
-
-Author: Brodie Mangan
-Version: 1.0
-"""
-
-# Licensed under the MIT License (see LICENSE file for full text)
+from wand_common import (create_grid, create_grid_lines, display_dual_stimulus,
+                         display_grid, draw_grid, generate_dual_nback_sequence,
+                         generate_positions_with_matches, get_jitter,
+                         get_level_color, get_param, get_text,
+                         install_error_hook, load_config, set_grid_lines)
 
 # --------------- CLI FLAGS (dummy‑run only) ---------------
-import argparse
-import random
 
 parser = argparse.ArgumentParser(add_help=False)
 parser.add_argument(
@@ -61,20 +75,20 @@ if getattr(sys, "frozen", False):
 else:
     base_dir = os.path.dirname(os.path.abspath(__file__))
 
+CONFIG_DIR = os.path.join(base_dir, "config")
+load_config(lang="en", config_dir=CONFIG_DIR)
 
-# === Experiment Settings ===
-FULLSCREEN = True  # Set to False for windowed mode during testing
-WINDOW_SIZE = (1200, 800)  # Only used if FULLSCREEN = False
-MONITOR_NAME = "testMonitor"  # Update to match lab monitor profile if necessary
-BACKGROUND_COLOR = "black"  # Background colour of experiment window
-COLOR_SPACE = "rgb"  # Colour space
-USE_FBO = True  # Framebuffer object for better rendering
-GRID_SPACING = 100  # Grid line spacing in pixels
-GRID_COLOR = "gray"  # Colour of grid lines
-GRID_OPACITY = 0.2  # Opacity of grid lines
-image_dir = os.path.join(
-    base_dir, "Abstract Stimuli", "apophysis"
-)  # Relative path to image folder
+# === Window configuration (from params.json) ===
+WIN_FULLSCR = bool(get_param("window.fullscreen", False))
+WIN_SIZE = tuple(get_param("window.size", [1650, 1000]))
+WIN_MONITOR = str(get_param("window.monitor", "testMonitor"))
+WIN_BG = get_param("window.background_color", "black")
+WIN_COLORSP = get_param("window.color_space", "rgb")
+WIN_USEFBO = bool(get_param("window.use_fbo", True))
+
[…]
