+++ WAND_practice_plateau.py
@@ -1,4 +1,30 @@
 #!/usr/bin/env python3
+"""
+WAND — Practice (Plateau Calibration)
+
+Practice session protocol for calibrating working-memory performance prior to
+cognitive-fatigue induction using the WAND model (Working-memory Adaptive-fatigue
+with N-back Difficulty).
+
+Participants complete Spatial, Dual, and Sequential N-back tasks with guided
+demos and brief feedback until performance stabilises.
+
+Author
+------
+Brodie E. Mangan
+
+Version
+-------
+1.0
+
+Environment
+-----------
+Tested on Windows, Python 3.8. See requirements.txt for exact pins.
+
+License
+-------
+MIT (see LICENSE).
+"""
 import argparse
 import math
 import os
@@ -8,23 +34,11 @@ import traceback
 
 from psychopy import core, event, visual
 
-"""
-WAND-practice-plateau
-
-Practice session protocol for calibrating working memory performance prior to cognitive fatigue induction
-using the WAND (Working-memory Adaptive-fatigue with N-back Difficulty) model.
-
-Participants complete Spatial, Dual, and Sequential N-back tasks with demonstrations until plateau is achieved. 
-Designed for pre-fatigue calibration in cognitive fatigue experiments, suitable
-for EEG and behavioural-only studies.
-
-Requires: PsychoPy, Python 3.8+.
-
-Author: Brodie Mangan
-Version: 1.0
-"""
-
-# Licensed under the MIT License (see LICENSE file for full text)
+from wand_common import (create_grid, create_grid_lines, display_dual_stimulus,
+                         display_grid, draw_grid, generate_dual_nback_sequence,
+                         generate_positions_with_matches, get_level_color,
+                         get_param, get_text, install_error_hook, load_config,
+                         set_grid_lines)
 
 if getattr(sys, "frozen", False):
     # if you’ve bundled into an executable
@@ -44,6 +58,17 @@ import datetime
 
 data_dir = os.path.join(base_dir, "data")
 os.makedirs(data_dir, exist_ok=True)
+
+CONFIG_DIR = os.path.join(base_dir, "config")
+load_config(lang="en", config_dir=CONFIG_DIR)
+
+# Window configuration from params.json
+WIN_FULLSCR = bool(get_param("window.fullscreen", False))
+WIN_SIZE = tuple(get_param("window.size", [1650, 1000]))
+WIN_MONITOR = str(get_param("window.monitor", "testMonitor"))
+WIN_BG = get_param("window.background_color", "black")
+WIN_COLORSP = get_param("window.color_space", "rgb")
+WIN_USEFBO = bool(get_param("window.use_fbo", True))
 
 
 def _prompt_participant_id(win) -> str:
@@ -75,9 +100,7 @@ def _prompt_participant_id(win) -> str:
     txt = dict(height=24, color="white", wrapWidth=900)
     pid = ""
     while True:
-        visual.TextStim(
-            win, text="Enter Participant ID then press Return", **txt, pos=(0, 120)
-        ).draw()
+        visual.TextStim(win, text=get_text("get_pid"), **txt, pos=(0, 120)).draw()
         visual.Rect(win, width=380, height=50, lineColor="white", pos=(0, 40)).draw()
         visual.TextStim(win, text=pid, **txt, pos=(0, 40)).draw()
         win.flip()
@@ -206,7 +229,18 @@ DISTRACTORS_ENABLED = None if args.distractors is None else (args.distractors !=
 
 
 def _apply_seed(seed_val):
-    """Seed Python & NumPy RNGs once, if a seed is provided."""
+    """
+    Seed Python and (optionally) NumPy RNGs once.
+
+    Parameters
[…]
