+++ Changelog.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [1.0.3] - 2025-10-03
+### Added
+- `wand_common.py` shared module centralising utilities (grid drawing, jitter, colour mapping, error hook, sequence generators). Both `WAND_full_induction.py` and `WAND_practice_plateau.py` import from this module.
+
+### Changed
+- **External configuration**: Most participant-facing text and key parameters are now in `config/text_en.json` and `config/params.json` to satisfy reviewer feedback and improve maintainability.
+- **Sequential N-back**: Clarified target generation to keep ~50% targets among eligible trials with a maximum of two consecutive matches.
+- **Sequential practice (slow mode)**: Slow-mode block length set to 60 trials to match other tasks.
+- **Docs**: README streamlined and task descriptions clarified.
+
+### Fixed
+- Minor logging wording in dual-task notes (no functional impact).
+
+### Removed
+- Small cleanup of unused variables / redundant lines across scripts for clarity.
+
+
+---
+
 ## [1.0.2] - 2025-06-12
 
 fixed - spatial background grid to not dissapear 
