### 2025-10-03 fbed516
style: Configure and apply black/isort formatting
