### 2025-10-03 #8667 @brodie-neuro
@mstimberg @OleBialas

Thank you both for your patience. Just to follow up, all the requested changes have now been completed and implemented.

I have addressed the feedback from @OleBialas and have just closed the two corresponding issues (#1 and #2) with comments and links to the changes.

A new version, v1.0.3, which incorporates all of this work, has now been published. Please let me know if there is anything else required from my end. Thank you again for your time and guidance.
