### 2025-10-07 #8667 @mstimberg
Many thanks for the update @brodie-neuro, I hope that addresses your concerns, @OleBialas, and you are able to continue your review.

@LeoDutriaux Could you please let us know whether you already had time to take a look at the software/paper and if not, when you estimate being able to do so 🙏
