### 2025-11-25 #8667 @mstimberg
Hi @brodie-neuro, hope you had a good annual leave! Can you give us an estimate when you will be able to address the open issues?
