### 2025-11-25 #8667 @brodie-neuro
Hi @mstimberg . I did, thank you very much. Funny enough, I just started yesterday. I would like to think that the comments will be addressed within the next week or so.
