### 2025-12-08 951f448
Release v1.0.4: Architecture refactor, plateau logic fix, docs update
