+++ WAND_practice_plateau.py
@@ -15,26 +15,38 @@ Brodie E. Mangan
 
 Version
 -------
-1.0
+1.0.4
 
 Environment
 -----------
 Tested on Windows, Python 3.8. See requirements.txt for exact pins.
 
+Configuration
+-------------
+See wand_common.py and the config/ folder for core experiment parameters,
+text strings, and window settings.
+
 License
 -------
 MIT (see LICENSE).
 """
+# =============================================================================
+#  SECTION 1: IMPORTS & SETUP
+# =============================================================================
 import argparse
+import csv
+import datetime
 import math
 import os
 import random
 import sys
 import traceback
+from typing import List, Tuple
 
 from psychopy import core, event, visual
 
 from wand_common import (
+    collect_trial_response,
     create_grid,
     create_grid_lines,
     display_dual_stimulus,
@@ -42,12 +54,16 @@ from wand_common import (
     draw_grid,
     generate_dual_nback_sequence,
     generate_positions_with_matches,
+    generate_sequential_image_sequence,
     get_level_color,
     get_param,
     get_text,
     install_error_hook,
     load_config,
+    prompt_choice,
+    prompt_text_input,
     set_grid_lines,
+    show_text_screen,
 )
 
 if getattr(sys, "frozen", False):
@@ -56,72 +72,44 @@ if getattr(sys, "frozen", False):
 else:
     base_dir = os.path.dirname(os.path.abspath(__file__))
 
-# ────────────────────────────────────────────────────────────────
-#  ▶▶  SEQUENTIAL BLOCK LOGGER  ◀◀
-#      • ask for Participant ID immediately
-#      • create ./data if missing
-#      • write rows:  PID,level,block,accuracy,lapses,errors
-#      • whenever the N-back level changes, start a new section
-# ────────────────────────────────────────────────────────────────
-import csv
-import datetime
-
 data_dir = os.path.join(base_dir, "data")
 os.makedirs(data_dir, exist_ok=True)
 
 CONFIG_DIR = os.path.join(base_dir, "config")
 load_config(lang="en", config_dir=CONFIG_DIR)
 
-# Window configuration from params.json
+# Window configuration
 WIN_FULLSCR = bool(get_param("window.fullscreen", False))
 WIN_SIZE = tuple(get_param("window.size", [1650, 1000]))
 WIN_MONITOR = str(get_param("window.monitor", "testMonitor"))
-WIN_BG = get_param("window.background_color", "black")
+WIN_BG = get_param("window.background_color", [-1, -1, -1])
 WIN_COLORSP = get_param("window.color_space", "rgb")
 WIN_USEFBO = bool(get_param("window.use_fbo", True))
 
-
+# =============================================================================
+#  SECTION 2: LOGGING & DATA MANAGEMENT
+# =============================================================================
 def _prompt_participant_id(win) -> str:
     """
-    Prompt the experimenter for a participant-ID via an on-screen textbox.
-
-    A minimal text editor is drawn in the centre of the PsychoPy window.
-    Characters are appended one-by-one; **Backspace** deletes the last
-    character.  The routine returns only when the user presses
-    **Return / Enter** *and* at least one character has been typed.
-
-    Parameters
-    ----------
-    win : psychopy.visual.Window
-        The active PsychoPy window.
+    Prompt the experimenter for a participant ID via an on-screen textbox.
 
     Returns
     -------
     str
-        A non-empty participant identifier.
-
-    Notes
-    -----
-    * The routine purposefully ignores *Escape* so the experiment cannot
-      proceed without an ID.
-    * The textbox width is fixed at 380 px; text wraps automatically
-      after 900 px to avoid runaway lines.
[…]
