+++ wand_common.py
@@ -1,18 +1,23 @@
 #!/usr/bin/env python3
 """
-wand_common.py
+WAND Common Utilities
 
-Shared helpers and configuration loader for the WAND tasks.
+Centralised library for the WAND (Working-memory Adaptive-fatigue with N-back Difficulty)
+suite. This module handles:
+  - Configuration loading (params.json, text_en.json)
+  - UI helpers (standardised prompts, text screens)
+  - Visual rendering (background grids, stimulus display)
+  - Core timing loops (response collection)
+  - Sequence generation algorithms (Spatial, Dual, Sequential)
 
-This module centralises:
-  - JSON-based configuration loading (parameters and UI text)
-  - Window and grid drawing helpers used across tasks
-  - Small utilities (timing jitter, colour selection, sequence generators)
-  - An error hook that renders a readable message in the PsychoPy window
+Author
+------
+Brodie E. Mangan
 
-It is safe to import from both the full induction and practice scripts.
+License
+-------
+MIT (see LICENSE).
 """
-
 from __future__ import annotations
 
 import inspect
@@ -21,16 +26,19 @@ import logging
 import os
 import random
 import sys
-from typing import Any, Dict, Iterable, List, Optional, Tuple
+from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Tuple
 
-from psychopy import core, visual  # type: ignore
+from psychopy import core, event, visual
 
-# -----------------------------------------------------------------------------
-# Module configuration and globals
-# -----------------------------------------------------------------------------
+# =============================================================================
+#  SECTION 1: CONFIGURATION & SYSTEM SETUP
+# =============================================================================
 
+# Global config cache
+_PARAMS = {}
+_TEXT = {}
+_GRID_LINES = []  # Cache for the static background grid lines
 LOGGER = logging.getLogger(__name__)
-
 # Base directory discovery (works when frozen with PyInstaller)
 if getattr(sys, "frozen", False):
     BASE_DIR = sys._MEIPASS  # type: ignore[attr-defined]
@@ -46,11 +54,6 @@ TEXT: Dict[str, str] = {}
 
 # Cached background grid lines
 _GRID_LINES: List[visual.ShapeStim] = []
-
-
-# =============================================================================
-# Configuration loading
-# =============================================================================
 
 
 def _safe_read_json(path: str) -> Any:
@@ -79,6 +82,32 @@ def _safe_read_json(path: str) -> Any:
     except Exception as exc:  # noqa: BLE001
         LOGGER.warning("Failed to parse JSON '%s': %s", path, exc)
     return None
+
+
+def _get_from(obj: Dict[str, Any], path: str, default: Any = None) -> Any:
+    """
+    Traverse a nested dictionary using a dotted key path.
+
+    Parameters
+    ----------
+    obj : Dict[str, Any]
+        Source dictionary.
+    path : str
+        Dotted path, for example "practice.speed_multiplier.slow".
+    default : Any, optional
+        Value returned when a key is missing. Default is None.
+
+    Returns
+    -------
+    Any
+        The resolved value or `default`.
+    """
+    cur: Any = obj
+    for part in path.split("."):
+        if not isinstance(cur, dict) or part not in cur:
+            return default
+        cur = cur[part]
+    return cur
 
 
 def load_config(
@@ -133,32 +162,6 @@ def load_config(
     return PARAMS, TEXT
 
 
-def _get_from(obj: Dict[str, Any], path: str, default: Any = None) -> Any:
-    """
-    Traverse a nested dictionary using a dotted key path.
-
-    Parameters
-    ----------
-    obj : Dict[str, Any]
-        Source dictionary.
-    path : str
-        Dotted path, for example "practice.speed_multiplier.slow".
-    default : Any, optional
[…]
