### 2025-12-08 5e60c4c
Finalize v1.0.4: Update all documentation, config, and tests
