+++ Tests/test_metrics.py
@@ -1,0 +1,264 @@
+import json
+import os
+import sys
+from math import isclose
+
+import pytest
+
+# Ensure we can import modules from the main folder
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+
+from wand_analysis import (
+    calculate_A_prime,
+    calculate_accuracy_and_rt,
+    calculate_dprime,
+    summarise_sequential_block,
+)
+
+# --- LOGGING HELPER ---
+# This allows us to save the "evidence" to a file
+LOG_FILE = "test_results_detailed.txt"
+
+
+def log_evidence(test_name, input_desc, expected, actual, status):
+    with open(LOG_FILE, "a") as f:
+        f.write(f"TEST: {test_name}\n")
+        f.write(f"  INPUT:    {input_desc}\n")
+        f.write(f"  EXPECTED: {expected}\n")
+        f.write(f"  ACTUAL:   {actual}\n")
+        f.write(f"  STATUS:   {status}\n")
+        f.write("-" * 50 + "\n")
+
+
+# Clear the log file at the start of the run
+@pytest.hookimpl(tryfirst=True)
+def pytest_sessionstart(session):
+    with open(LOG_FILE, "w") as f:
+        f.write("=== WAND BEHAVIOURAL METRICS TEST EVIDENCE ===\n\n")
+
+
+# --- TEST DATA FIXTURES ---
+
+
+@pytest.fixture
+def perfect_data():
+    """Returns a list of trials where the user got everything 100% right."""
+    return [
+        {
+            "Trial": 1,
+            "Is Target": True,
+            "Response": "match",
+            "Reaction Time": 0.5,
+            "Accuracy": True,
+        },
+        {
+            "Trial": 2,
+            "Is Target": True,
+            "Response": "match",
+            "Reaction Time": 0.5,
+            "Accuracy": True,
+        },
+        {
+            "Trial": 3,
+            "Is Target": False,
+            "Response": "non-match",
+            "Reaction Time": 0.5,
+            "Accuracy": True,
+        },
+        {
+            "Trial": 4,
+            "Is Target": False,
+            "Response": "non-match",
+            "Reaction Time": 0.5,
+            "Accuracy": True,
+        },
+    ]
+
+
+@pytest.fixture
+def random_data():
+    """Returns data behaving like a random guesser (50% accuracy)."""
+    return [
+        {
+            "Trial": 1,
+            "Is Target": True,
+            "Response": "match",
+            "Reaction Time": 0.5,
+            "Accuracy": True,
+        },  # Hit
+        {
+            "Trial": 2,
+            "Is Target": True,
+            "Response": "non-match",
+            "Reaction Time": 0.5,
+            "Accuracy": False,
+        },  # Miss
+        {
+            "Trial": 3,
+            "Is Target": False,
+            "Response": "match",
+            "Reaction Time": 0.5,
+            "Accuracy": False,
+        },  # False Alarm
+        {
+            "Trial": 4,
+            "Is Target": False,
+            "Response": "non-match",
+            "Reaction Time": 0.5,
+            "Accuracy": True,
+        },
+        # Correct Rejection
+    ]
+
+
+@pytest.fixture
+def block_data_with_distractor():
+    """
+    Constructs a 10-trial block with a distractor at Trial 5.
+
+    Structure:
+    - Trials 1-4: Perfect performance (Pre-Distractor)
[…]
