### 2025-12-08 7795cb1
Fix CI: Install requirements.txt to enable scipy tests
