+++ .github/workflows/ci.yml
@@ -26,10 +26,12 @@ jobs:
         with:
           python-version: '3.8'
 
-      - name: Install testing tools
+      - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install pytest pytest-mock black==22.10.0 isort==5.12.0
+          # Install project requirements (Scipy, PsychoPy, etc.) if file exists
+          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
       - name: Run black (check only)
         run: black --check .
