### 2025-12-08 088dbbd
Switch CI to Windows runner to support pywin32 dependency
