+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ on:
 jobs:
   test:
     name: Lint and Test
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
 
     steps:
       - name: Checkout code
@@ -27,17 +27,21 @@ jobs:
           python-version: '3.8'
 
       - name: Install dependencies
+        shell: bash
         run: |
           python -m pip install --upgrade pip
           pip install pytest pytest-mock black==22.10.0 isort==5.12.0
-          # Install project requirements (Scipy, PsychoPy, etc.) if file exists
+          # Install project requirements (Scipy, PsychoPy, pywin32, etc.)
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
       - name: Run black (check only)
+        shell: bash
         run: black --check .
 
       - name: Run isort (check only)
+        shell: bash
         run: isort --check .
-      
+
       - name: Run tests
+        shell: bash
         run: pytest --maxfail=1 --disable-warnings -q
