### 2026-01-05 79e7274
CI: Verbose pytest output; ignore local monitor test script
