+++ .github/workflows/ci.yml
@@ -44,4 +44,4 @@ jobs:
 
       - name: Run tests
         shell: bash
-        run: pytest --maxfail=1 --disable-warnings -q
+        run:  pytest -v --tb=short
