### 2026-01-06 484bc34
Add --quicktest flag for automated smoke testing
