+++ Dummy_Run.py
@@ -11,18 +11,104 @@ Standalone runner for a short Sequential N-back check using the WAND suite.
 
 Use this script to verify task logic, display settings, and data output.
 
+Quicktest Mode (--quicktest):
+- Mocks keyboard input (no human interaction required)
+- Uses fast timings (0.05s display, 0.05s ISI)
+- Runs 10 trials in ~5 seconds
+- Verifies PsychoPy installation works without crashing
+
 Author: Brodie Mangan
-Version: 1.0.0
+Version: 1.1.0
 License: MIT
 Requires: PsychoPy ≥ 2024.2.1, Python 3.8+
 """
 
 import argparse
 import os
+import random
 import sys
 import time
 
-# Import required functions and constants from the main induction script
+# =============================================================================
+# ARGUMENT PARSING (must happen before conditional mocking)
+# =============================================================================
+
+parser = argparse.ArgumentParser(description="Quick dummy Sequential N‑back test")
+parser.add_argument(
+    "--level", type=int, default=2, help="N‑back level (default = 2)"
+)
+parser.add_argument(
+    "--trials", type=int, default=35, help="Number of trials (default = 35)"
+)
+parser.add_argument(
+    "--quicktest",
+    action="store_true",
+    help="Run automated smoke test (no human input required, fast timings)",
+)
+args = parser.parse_args()
+
+# =============================================================================
+# CONDITIONAL MOCKING FOR QUICKTEST MODE
+# =============================================================================
+# This MUST happen before importing WAND modules which import psychopy
+
+QUICKTEST_DISPLAY = 0.05
+QUICKTEST_ISI = 0.05
+
+if args.quicktest:
+    print("[QUICKTEST] Automated smoke test mode enabled")
+    print("[QUICKTEST] Patching PsychoPy event functions...")
+
+    from psychopy import event, core
+
+    _original_waitKeys = event.waitKeys
+    _original_getKeys = event.getKeys
+
+    def mock_waitKeys(keyList=None, maxWait=float('inf'), *args, **kwargs):
+        """
+        Mock replacement for event.waitKeys.
+        Returns a random key from keyList immediately without waiting.
+        """
+        core.wait(0.001)
+
+        if keyList is None:
+            keyList = ['space']
+
+        safe_keys = [k for k in keyList if k != 'escape']
+        if not safe_keys:
+            safe_keys = ['space']
+
+        return [random.choice(safe_keys)]
+
+    def mock_getKeys(keyList=None, *args, **kwargs):
+        """
+        Mock replacement for event.getKeys.
+        Simulates ~40% chance of a keypress (realistic response rate).
+        """
+        if keyList is None:
+            return []
+
+        safe_keys = [k for k in keyList if k != 'escape']
+        if not safe_keys:
+            return []
+
+        if random.random() < 0.4:
+            return [random.choice(safe_keys)]
+        return []
+
+    event.waitKeys = mock_waitKeys
+    event.getKeys = mock_getKeys
+
+    random.seed(12345)
+    args.trials = 10
+
+    print("[QUICKTEST] PsychoPy event functions patched successfully")
+    print(f"[QUICKTEST] Using {args.trials} trials with {QUICKTEST_DISPLAY}s display, {QUICKTEST_ISI}s ISI")
+
+# =============================================================================
+# IMPORT WAND MODULES (after potential mocking)
+# =============================================================================
+
 from WAND_full_induction import (
     base_dir,
[…]
