### 2026-01-13 4b1c42e
v1.1.0: pip-installable package, GUI Launcher, Block Builder, EEG triggers, full SDT metrics
