+++ WAND_Launcher.py
@@ -1,0 +1,1381 @@
+#!/usr/bin/env python3
+"""
+WAND Launcher v1.1.0 - Comprehensive Study Configuration System
+
+A professional GUI for configuring WAND experiments with:
+- Study presets (save/load configurations)
+- Per-task timing settings
+- Task enable/disable toggles
+- Time compression controls
+- Counterbalancing options
+
+Author: Brodie E. Mangan
+Version: 1.1.0
+License: MIT
+"""
+
+import json
+import os
+import sys
+from collections import OrderedDict
+from datetime import datetime
+
+# Import block builder module
+try:
+    from block_builder import show_block_builder
+
+    BLOCK_BUILDER_AVAILABLE = True
+except ImportError:
+    BLOCK_BUILDER_AVAILABLE = False
+    print("[WARNING] block_builder.py not found - block ordering disabled")
+
+# =============================================================================
+# Import PsychoPy GUI
+# =============================================================================
+# Scale Qt dialogs to be 20% larger (set before importing psychopy.gui)
+os.environ["QT_SCALE_FACTOR"] = "1.2"
+
+try:
+    from psychopy import gui, logging
+
+    # Suppress cosmetic validation errors from GUI (e.g. "could not convert string to float")
+    logging.console.setLevel(logging.CRITICAL)
+except ImportError:
+    print("ERROR: PsychoPy is not installed.")
+    print("Please run: pip install -r requirements.txt")
+    sys.exit(1)
+
+# =============================================================================
+# Paths and Directories
+# =============================================================================
+BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+CONFIG_DIR = os.path.join(BASE_DIR, "config")
+PRESETS_DIR = os.path.join(CONFIG_DIR, "presets")
+DATA_DIR = os.path.join(BASE_DIR, "data")
+
+# Ensure directories exist
+os.makedirs(PRESETS_DIR, exist_ok=True)
+os.makedirs(DATA_DIR, exist_ok=True)
+
+# =============================================================================
+# Default Configuration
+# =============================================================================
+DEFAULT_CONFIG = {
+    # Metadata
+    "study_name": "Default_Study",
+    "created": None,
+    "last_modified": None,
+    # Task Mode
+    "task_mode": "Full Induction",
+    # Task Enables
+    "sequential_enabled": True,
+    "spatial_enabled": True,
+    "dual_enabled": True,
+    # Sequential N-back Settings
+    "sequential": {
+        "blocks": 5,
+        "display_duration": 0.8,
+        "isi": 1.0,
+        "trials_per_block": 164,
+        "distractors_enabled": True,
+    },
+    # Spatial N-back Settings
+    "spatial": {
+        "blocks": 4,
+        "display_duration": 1.0,
+        "isi": 1.0,
+        "time_compression": True,
+    },
+    # Dual N-back Settings
+    "dual": {
+        "blocks": 4,
+        "display_duration": 1.0,
+        "isi": 1.2,
+        "time_compression": True,
+    },
+    # Global Options
+    "counterbalance_spatial_dual": False,
+    "fullscreen": True,
+    "rng_seed": None,
+    "breaks_schedule": [2, 4],
+    "measures_schedule": [2, 3, 4, 5],
+}
+
+
+# =============================================================================
+# Preset Management
+# =============================================================================
+
+
+def get_available_presets():
+    """
+    Get list of available study presets.
+
+    Returns
+    -------
+    list
+        List of preset names (without .json extension)
+    """
[…]
