### 2026-01-23 eb4b5b5
v1.1.1: Tooltips, duration fixes, reviewer feedback
