+++ wand_nback/launcher.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-WAND Launcher v1.1.0 - Comprehensive Study Configuration System
+WAND Launcher v1.1.1 - Comprehensive Study Configuration System
 
 A professional GUI for configuring WAND experiments with:
 - Study presets (save/load configurations)
@@ -10,7 +10,7 @@ A professional GUI for configuring WAND experiments with:
 - Counterbalancing options
 
 Author: Brodie E. Mangan
-Version: 1.1.0
+Version: 1.1.1
 License: MIT
 """
 
@@ -22,7 +22,7 @@ from datetime import datetime
 
 # Import block builder module
 try:
-    from block_builder import show_block_builder
+    from wand_nback.block_builder import show_block_builder
 
     BLOCK_BUILDER_AVAILABLE = True
 except ImportError:
@@ -100,6 +100,28 @@ DEFAULT_CONFIG = {
     "breaks_schedule": [2, 4],
     "measures_schedule": [2, 3, 4, 5],
 }
+
+
+# =============================================================================
+# Utility Functions
+# =============================================================================
+
+
+def format_duration(minutes):
+    """Format duration as 'X min Y sec' for better resolution on short experiments."""
+    if minutes >= 60:
+        hours = int(minutes // 60)
+        mins = int(minutes % 60)
+        return f"{hours}h {mins}m"
+    elif minutes >= 1:
+        mins = int(minutes)
+        secs = int((minutes - mins) * 60)
+        if secs > 0:
+            return f"{mins} min {secs} sec"
+        return f"{mins} min"
+    else:
+        secs = int(minutes * 60)
+        return f"{secs} sec"
 
 
 # =============================================================================
@@ -204,10 +226,10 @@ def show_page1_study_setup():
     """
     presets = get_available_presets()
 
-    # Ensure "Default" is always available and first after <Create New>
-    if "Default" in presets:
-        presets.remove("Default")
-    presets.insert(0, "Default")  # Put Default FIRST (before <Create New>)
+    # Ensure "Standard_WAND_Protocol" is always available and first after <Create New>
+    if "Standard_WAND_Protocol" in presets:
+        presets.remove("Standard_WAND_Protocol")
+    presets.insert(0, "Standard_WAND_Protocol")  # Put Standard first
     # Move <Create New> to end of list
     if "<Create New>" in presets:
         presets.remove("<Create New>")
@@ -215,17 +237,25 @@ def show_page1_study_setup():
 
     # Use OrderedDict with DlgFromDict for reliable cross-version behaviour
     fields = OrderedDict()
-    fields["Load_Preset"] = presets  # Dropdown - Default is now first
+    fields["Load_Preset"] = presets  # Dropdown - Standard_WAND_Protocol is now first
     fields["Study_Name"] = "My_Study"
     fields[
         "Participant_ID"
     ] = ""  # Required but without asterisk to avoid PsychoPy message
+
+    # Tooltips for Study Setup
+    tips = {
+        "Load_Preset": "Load a saved configuration. Standard_WAND_Protocol uses default parameters.",
+        "Study_Name": "Name of your study. Used in output file names.",
+        "Participant_ID": "Unique participant identifier. Used in output file names.",
+    }
 
     dlg = gui.DlgFromDict(
         dictionary=fields,
         title="WAND - Page 1/5: Study Setup",
         sortKeys=False,
         show=False,
+        tip=tips,
     )
 
     dlg.show()
@@ -281,10 +311,21 @@ def show_page2_task_selection(config):
     fields["Dual_Enabled"] = config.get("dual_enabled", True)
     fields["Dual_Blocks"] = config.get("dual", {}).get("blocks", 4)
 
[…]
