### 2026-01-23 6cda3c8
v1.1.1: Paper updates and code formatting
