+++ Tests/test_imports.py
@@ -16,44 +16,50 @@ import pytest
 def test_import_analysis():
     """Verify wand_nback.analysis can be imported."""
     from wand_nback import analysis
-    assert hasattr(analysis, 'calculate_dprime')
-    assert hasattr(analysis, 'calculate_sdt_metrics')
+
+    assert hasattr(analysis, "calculate_dprime")
+    assert hasattr(analysis, "calculate_sdt_metrics")
 
 
 def test_import_common():
     """Verify wand_nback.common can be imported."""
     from wand_nback import common
-    assert hasattr(common, 'load_gui_config')
+
+    assert hasattr(common, "load_gui_config")
 
 
 def test_import_block_builder():
     """Verify wand_nback.block_builder can be imported."""
     from wand_nback import block_builder
-    assert hasattr(block_builder, 'show_block_builder')
+
+    assert hasattr(block_builder, "show_block_builder")
 
 
 def test_import_launcher():
     """Verify wand_nback.launcher can be imported."""
     from wand_nback import launcher
-    assert hasattr(launcher, 'main')
+
+    assert hasattr(launcher, "main")
 
 
 def test_import_practice_plateau():
     """Verify wand_nback.practice_plateau can be imported."""
     from wand_nback import practice_plateau
-    assert hasattr(practice_plateau, 'main')
+
+    assert hasattr(practice_plateau, "main")
 
 
 def test_import_full_induction():
     """Verify wand_nback.full_induction can be imported.
-    
+
     Note: This may fail in test environments due to PsychoPy initialization.
     The actual runtime import works correctly.
     """
     try:
         from wand_nback import full_induction
-        assert hasattr(full_induction, 'main_task_flow')
+
+        assert hasattr(full_induction, "main_task_flow")
     except (AttributeError, ImportError) as e:
         # PsychoPy mock issues during testing - skip but warn
-        if '__version__' in str(e) or 'psychopy' in str(e).lower():
+        if "__version__" in str(e) or "psychopy" in str(e).lower():
             pytest.skip("PsychoPy initialization issue in test environment")
