### 2026-01-23 37a5b8e
Fix CI headless tests: Mock PsychoPy/Pyglet in conftest.py
