### 2026-01-23 dc6696b
Style: Format conftest.py with black
