### 2026-01-27 #8667 @OleBialas
Hey @mstimberg,
The project has come a long way. @brodie-neuro has reorganized the code so that it can be installed as a pip package. He also added an automated test that users can run to assert the program is working, as well as a GUI to configure the experiment.
I just opened a few very minor issues on adding docs and improving the appearance and once they are addressed I would consider the software ready for publishing
