### 2026-04-14 #8667 @mstimberg
Hi @OleBialas It looks as if @brodie-neuro addressed all the remaining minor issues you raised – could you please confirm that this addresses all your concerns and that you recommend the software/paper for publication?

@LeoDutriaux Could you please get back to us about your review?
