### 2026-04-16 #8667 @LeoDutriaux
I've been through the whole software (and I guess it works, I'm fatigued ahah) and it's honestly very cool! The GUI is, in particular, very well made and convenient.
I raised a couple of issues, and after they are addressed, I definitely recommend the software/paper for publication.
