### 2026-04-16 #8667 @OleBialas
> Hi [@OleBialas](https://github.com/OleBialas) It looks as if [@brodie-neuro](https://github.com/brodie-neuro) addressed all the remaining minor issues you raised – could you please confirm that this addresses all your concerns and that you recommend the software/paper for publication?
>
> [@LeoDutriaux](https://github.com/LeoDutriaux) Could you please get back to us about your review?

All good from my side!
