### 2026-04-24 #8667 @brodie-neuro
Hi @mstimberg, @OleBialas, and @LeoDutriaux,

I've pushed the reviewer changes and made a new release (v1.3.2).

A quick note on the review title: "Active-fatigue" was a working title. The acronym is now "Adaptive-fatigue," which is consistent across the paper, README, and all metadata.

Thank you all for the thorough review process, it has been much appreciated.
