### 2026-04-29 #8667 @mstimberg
Thanks for the update @brodie-neuro Note that the title of the final paper will be picked up from the `paper.md` file, so the change shouldn't be an issue. To avoid confusion, I will also update the title of this review issue.

@LeoDutriaux Could you please give us a quick confirmation that your last remarks have been addressed with the latest changes and that everything looks good for publication from your side?
