### 2026-04-30 #8667 @mstimberg
(Reposting my comment since the list overwrote it)
With both reviewers recommending acceptance, I will now proceed with some final checks before handing the paper over to the topic editor. @brodie-neuro Could you please take care of the "author tasks" in the list ☝️  and post the information here as a comment? Thanks!
