### 2026-04-30 #8667 @mstimberg
@brodie-neuro Could you please fix the DOI mentioned above (I think it should be dots instead of underscores)?

Also, for this reference:
```bibtex
@unpublished{mangan2026validation,
  title={Validating the {WAND} fatigue induction protocol: Behavioural evidence for active cognitive fatigue},
  author={Mangan, Brodie E. and Kourtis, Dimitrios and Tomaz, Simone},
  year={2026},
  note={Manuscript under review}
}
```
Wouldn't it make sense to replace it with the preprint at https://doi.org/10.31234/osf.io/h4n2w_v1 ?
Please note that from JOSS's side, any changes that are only to the paper/bibliography do not require you do make a new release of the software.
