### 2026-05-10 #8667 @brodie-neuro
Thanks @mstimberg. Apologies for the slight delay. PR #19 has been merged. Here is the requested information:

Version: v1.3.2 Archive DOI: https://doi.org/10.5281/zenodo.19719398

The Zenodo archive title, author list (including ORCID), and license (MIT) all match the JOSS paper. No software changes have been made since the tagged release, only paper/bibliography edits.
